Use EditViewImplEx in desktop templates
