+++ VaadinatorTooling/VaadinatorTemplatesTouchkit/src/main/resources/de/akquinet/engineering/vaadinator/templates/AddViewImpl.template
@@ -25,8 +25,12 @@ import ${presenterPackage}.${bean.className}AddPresenter;
 import ${presenterPackage}.${bean.className}EditPresenter;
 import ${common.basePackage}.ui.view.ExceptionMappingStrategy;
 
-public class ${bean.className}AddViewImpl extends ${bean.className}EditViewImpl<${bean.className}AddView.Observer> implements ${bean.className}AddView {
-
+public class ${bean.className}AddViewImpl extends 
+#if( $bean.getDisplayProfile($profileName).getExStrings().contains("EDIT_VIEW_IMPL") )
+    ${bean.className}EditViewImplEx<${bean.className}AddView.Observer> implements ${bean.className}AddView {
+#else
+    ${bean.className}EditViewImpl<${bean.className}AddView.Observer> implements ${bean.className}AddView {
+#end
 	/**
 	 * 
 	 */
