+++ VaadinatorTooling/VaadinatorTemplatesTouchkit/src/main/resources/de/akquinet/engineering/vaadinator/templates/ChangeViewImpl.template
@@ -25,8 +25,12 @@ import ${presenterPackage}.${bean.className}ChangePresenter;
 import ${presenterPackage}.${bean.className}EditPresenter;
 import ${common.basePackage}.ui.view.ExceptionMappingStrategy;
 
-public class ${bean.className}ChangeViewImpl extends ${bean.className}EditViewImpl<${bean.className}ChangeView.Observer> implements	${bean.className}ChangeView  {
-
+public class ${bean.className}ChangeViewImpl extends
+#if( $bean.getDisplayProfile($profileName).getExStrings().contains("EDIT_VIEW_IMPL") ) 
+    ${bean.className}EditViewImplEx<${bean.className}ChangeView.Observer> implements	${bean.className}ChangeView  {
+#else
+    ${bean.className}EditViewImpl<${bean.className}ChangeView.Observer> implements	${bean.className}ChangeView  {
+#end
 	/**
 	 * 
 	 */
